Menu Functionality/Volume Functionality
Fixed the buttons in the Settings and Credits pages so now all of the buttons work and can navigate between them.
Added a SettingsController script to controller two sliders, one for Game Volume and one for SFX Volume. The sliders update the text field to their right, but do not affect the game yet as there is no audio yet